@@ -2,31 +2,28 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
 class MyBottomNav extends StatelessWidget {
-  MyBottomNav({
-    Key key,
-    this.height = 50.0,
-    @required this.items,
-    this.currentIndex = 0,
-    this.onTap,
-    this.iconSize,
-    this.color = CupertinoColors.inactiveGray,
-    this.activeColor,
-    this.backgroundColor = const Color(0xfff9f9f9),
-    this.border = const Border(
-      top: BorderSide(
+  MyBottomNav(
+      {Key? key,
+      this.height = 50.0,
+      required this.items,
+      this.currentIndex = 0,
+      this.onChange,
+      this.iconSize = 23,
+      this.color = CupertinoColors.inactiveGray,
+      this.activeColor,
+      this.backgroundColor = const Color(0xfff9f9f9),
+      this.border = const Border(
+          top: BorderSide(
         width: 0.4,
         color: Color(0x33000000),
         style: BorderStyle.solid,
-      )
-    )
-  }) :assert(items != null),
-      assert(
-        items.length >= 2,
-        "Tabs need at least 2 items to conform to Apple's HIG",
-      ),
-      assert(currentIndex != null),
-      assert(0 <= currentIndex && currentIndex < items.length),
-      super(key: key);
+      ))})
+      : assert(
+          items.length >= 2,
+          "Tabs need at least 2 items to conform to Apple's HIG",
+        ),
+        assert(0 <= currentIndex && currentIndex < items.length),
+        super(key: key);
 
   final double height;
 
@@ -34,9 +31,9 @@ class MyBottomNav extends StatelessWidget {
 
   final int currentIndex;
 
-  final ValueChanged<int> onTap;
+  final ValueChanged<int>? onChange;
 
-  final Color activeColor;
+  final Color? activeColor;
 
   final Color color;
 
@@ -50,7 +47,7 @@ class MyBottomNav extends StatelessWidget {
   Widget build(BuildContext context) {
     final double bottomPadding = MediaQuery.of(context).padding.bottom;
     return Container(
-      height: height+bottomPadding,
+      height: height + bottomPadding,
       padding: EdgeInsets.only(bottom: bottomPadding),
       decoration: BoxDecoration(
         border: border,
@@ -71,60 +68,68 @@ class MyBottomNav extends StatelessWidget {
       result.add(Expanded(
         child: GestureDetector(
           behavior: HitTestBehavior.opaque,
-          onTap: onTap == null ? null : () { onTap(index); },
+          onTap: () {
+            if (onChange != null) onChange!(index);
+          },
           child: Center(
             child: Stack(
-              overflow: Overflow.visible,
+              clipBehavior: Clip.none,
               children: <Widget>[
                 Column(
                   mainAxisAlignment: MainAxisAlignment.center,
                   children: <Widget>[
-                    isActive?
-                    item.activeIcon==null?IconTheme(
-                      data: IconThemeData(color: activeColor??Theme.of(context).primaryColor, size: iconSize),
-                      child: item.icon,
-                    ):item.activeIcon:
-                    IconTheme(
-                      data: IconThemeData(color: color, size: iconSize),
-                      child: item.icon,
-                    ),
+                    if (item.icon != null)
+                      IconTheme(
+                        data: IconThemeData(
+                            color: isActive
+                                ? (activeColor ??
+                                    Theme.of(context).primaryColor)
+                                : color,
+                            size: iconSize),
+                        child: item.activeIcon ?? item.icon!,
+                      ),
                     Text(
                       item.title,
                       style: TextStyle(
                         fontSize: 12,
-                        color: isActive?activeColor??Theme.of(context).primaryColor:color,
+                        color: isActive
+                            ? activeColor ?? Theme.of(context).primaryColor
+                            : color,
                       ),
                     ),
                   ],
                 ),
-                item.badge!=null?Positioned(
-                  top: 2,
-                  right: -5,
-                  child: item.badge>0?Container(
-                    width: 15,
-                    height: 15,
-                    decoration: BoxDecoration(
-                      shape: BoxShape.circle,
-                      color: Colors.red,
-                    ),
-                    child: Center(
-                      child: Text(
-                        item.badge.toString(),
-                        style: TextStyle(
-                          color: Colors.white,
-                          fontSize: 10,
-                        ),
-                      ),
-                    ),
-                  ):Container(
-                    width: 10,
-                    height: 10,
-                    decoration: BoxDecoration(
-                      shape: BoxShape.circle,
-                      color: Colors.red,
-                    ),
-                  ),
-                ):SizedBox.shrink(),
+                if (item.badge != null)
+                  Positioned(
+                    top: 2,
+                    right: -5,
+                    child: item.badge! > 0
+                        ? Container(
+                            width: 15,
+                            height: 15,
+                            decoration: BoxDecoration(
+                              shape: BoxShape.circle,
+                              color: Colors.red,
+                            ),
+                            child: Center(
+                              child: Text(
+                                item.badge.toString(),
+                                style: TextStyle(
+                                  color: Colors.white,
+                                  fontSize: 10,
+                                ),
+                              ),
+                            ),
+                          )
+                        : Container(
+                            width: 10,
+                            height: 10,
+                            decoration: BoxDecoration(
+                              shape: BoxShape.circle,
+                              color: Colors.red,
+                            ),
+                          ),
+                  )
               ],
             ),
           ),
@@ -137,17 +142,17 @@ class MyBottomNav extends StatelessWidget {
 
 class MyBottomNavItem {
   MyBottomNavItem({
-    @required this.icon,
+    this.icon,
     this.activeIcon,
-    this.title,
+    required this.title,
     this.badge,
-  }) : assert(icon != null);
+  });
 
-  final Icon icon;
+  final Icon? icon;
 
-  final Icon activeIcon;
+  final Icon? activeIcon;
 
   final String title;
 
-  final int badge;
+  final int? badge;
 }
