@@ -6,9 +6,8 @@ import 'package:native_pdf_view/native_pdf_view.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:uuid/uuid.dart';
 
-import '../../service.dart';
+import '../../service/service.dart';
 import '../../variable.dart';
-
 
 class PDFView extends StatefulWidget {
   static const name = '/pdfView';
@@ -22,7 +21,7 @@ class PDFView extends StatefulWidget {
 }
 
 class PDFViewState extends State<PDFView> {
-  PdfController _pdfController;
+  PdfController? _pdfController;
   int currentPage = 0;
   int allPage = 0;
 
@@ -36,46 +35,45 @@ class PDFViewState extends State<PDFView> {
           title: Text('PDF View'),
           actions: [
             Center(
-              child: Padding(
-                padding: EdgeInsets.only(right: width(30)),
-                child: Text(
-                  '$currentPage/$allPage',
-                  style: TextStyle(
-                    fontSize: 20,
-                    fontWeight: FontWeight.w600
-                  ),
-                ),
-              )
-            )
+                child: Padding(
+              padding: EdgeInsets.only(right: width(30)),
+              child: Text(
+                '$currentPage/$allPage',
+                style: TextStyle(fontSize: 20, fontWeight: FontWeight.w600),
+              ),
+            ))
           ],
         ),
       ),
-      body: _pdfController!=null?PdfView(
-        scrollDirection: Axis.vertical,
-        documentLoader: Center(child: CircularProgressIndicator()),
-        pageLoader: Center(child: CircularProgressIndicator()),
-        controller: _pdfController,
-        onDocumentLoaded: (PdfDocument document) {
-          setState(() {
-            currentPage = 1;
-            allPage = document.pagesCount;
-          });
-        },
-        onDocumentError: (error) {
-          print(error);
-        },
-        onPageChanged: (int page) {
-          setState(() {
-            currentPage = page;
-          });
-        },
-      ):Container(),
+      body: _pdfController != null
+          ? PdfView(
+              scrollDirection: Axis.vertical,
+              documentLoader: Center(child: CircularProgressIndicator()),
+              pageLoader: Center(child: CircularProgressIndicator()),
+              controller: _pdfController!,
+              onDocumentLoaded: (PdfDocument document) {
+                setState(() {
+                  currentPage = 1;
+                  allPage = document.pagesCount;
+                });
+              },
+              onDocumentError: (error) {
+                print(error);
+              },
+              onPageChanged: (int page) {
+                setState(() {
+                  currentPage = page;
+                });
+              },
+            )
+          : Container(),
     );
   }
 
   void _urlChange() async {
-    Directory tempDir = Directory.systemTemp ?? await getTemporaryDirectory();
-    String tempPath = tempDir.path + '/' + Uuid().v4() + '.pdf'; // /tmp ?? /Library/Caches
+    Directory tempDir = Directory.systemTemp;
+    String tempPath =
+        tempDir.path + '/' + Uuid().v4() + '.pdf'; // /tmp ?? /Library/Caches
     try {
       await Service.download(widget.url, tempPath);
     } catch (e) {
@@ -84,9 +82,7 @@ class PDFViewState extends State<PDFView> {
     _pdfController = PdfController(
       document: PdfDocument.openFile(tempPath),
     );
-    setState(() {
-
-    });
+    setState(() {});
   }
 
   @override
