@@ -16,23 +16,30 @@ class MyWebView extends StatefulWidget {
 }
 
 class _WebViewState extends State<MyWebView> {
-  InAppWebViewController controller;
+  InAppWebViewController? controller;
   int _progress = 0;
   String _windowTitle = 'WebView';
+  StateSetter titleStateSetter = (VoidCallback fn) {};
+  StateSetter progressStateSetter = (VoidCallback fn) {};
 
   get url => widget.url;
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar:  AppBar(
+      appBar: AppBar(
         centerTitle: true,
         toolbarHeight: 50,
-        title: Text(_windowTitle),
+        title: StatefulBuilder(
+          builder: (context, _setState) {
+            titleStateSetter = _setState;
+            return Text(_windowTitle);
+          },
+        ),
         leading: BackButton(
           onPressed: () async {
-            if (controller!=null && await controller.canGoBack()) {
-              controller.goBack();
+            if (controller != null && await controller!.canGoBack()) {
+              controller!.goBack();
             } else {
               Navigator.of(context).pop();
             }
@@ -47,22 +54,25 @@ class _WebViewState extends State<MyWebView> {
             child: Stack(
               children: <Widget>[
                 Container(
-                  color: Colors.indigo,
+                  color: Colors.white,
                 ),
                 InAppWebView(
-                  initialUrl: url,
+                  initialUrlRequest: URLRequest(url: Uri.parse(url)),
                   initialOptions: InAppWebViewGroupOptions(
                     crossPlatform: InAppWebViewOptions(
                       supportZoom: false,
                       mediaPlaybackRequiresUserGesture: false,
+                      transparentBackground: true,
                       // javaScriptCanOpenWindowsAutomatically: true,
                     ),
                     android: AndroidInAppWebViewOptions(
-                      verticalScrollbarPosition: AndroidVerticalScrollbarPosition.SCROLLBAR_POSITION_RIGHT,
-                      cacheMode: AndroidCacheMode.LOAD_CACHE_ELSE_NETWORK
+                      verticalScrollbarPosition:
+                          AndroidVerticalScrollbarPosition
+                              .SCROLLBAR_POSITION_RIGHT,
+                      cacheMode: AndroidCacheMode.LOAD_CACHE_ELSE_NETWORK,
                     ),
                     ios: IOSInAppWebViewOptions(
-                      // disallowOverScroll: false,
+                      disallowOverScroll: false,
                       allowsInlineMediaPlayback: true,
                       alwaysBounceVertical: true,
                     ),
@@ -70,42 +80,56 @@ class _WebViewState extends State<MyWebView> {
                   onWebViewCreated: (InAppWebViewController c) {
                     controller = c;
                     // flutter和webview通信
-                    controller.addJavaScriptHandler(handlerName: 'push', callback: (arguments) {
-                      Navigator.of(context).pushNamed('${MyWebView.name}?url=${Uri.encodeQueryComponent(arguments.first)}');
-                    });
+                    controller!.addJavaScriptHandler(
+                        handlerName: 'push',
+                        callback: (arguments) {
+                          Navigator.of(context).pushNamed(
+                              '${MyWebView.name}?url=${Uri.encodeQueryComponent(arguments.first)}');
+                        });
                   },
-                  onCreateWindow: (InAppWebViewController controller, CreateWindowRequest request) async {
-                    controller.loadUrl(url: request.url);
+                  onCreateWindow: (InAppWebViewController controller,
+                      CreateWindowAction request) async {
+                    controller.loadUrl(
+                        urlRequest: URLRequest(url: request.request.url));
                     return false;
                   },
-                  onProgressChanged: (InAppWebViewController controller, int progress) {
-                    setState(() {
+                  onProgressChanged:
+                      (InAppWebViewController controller, int progress) {
+                    progressStateSetter(() {
                       _progress = progress;
                     });
                   },
-                  onTitleChanged: (InAppWebViewController controller, String title) {
-                    setState(() {
-                      _windowTitle = title;
-                    });
+                  onTitleChanged:
+                      (InAppWebViewController controller, String? title) {
+                    if (title != null) {
+                      titleStateSetter(() {
+                        _windowTitle = title;
+                      });
+                    }
                   },
                   onConsoleMessage: (controller, consoleMessage) {
                     print("console: " + consoleMessage.message);
                   },
                 ),
-                _progress!=100?Positioned(
-                  child: GestureDetector(
-                    behavior: HitTestBehavior.opaque,
-                    child: Center(
-                      child: Container(
-                        width: width(100),
-                        height: width(100),
-                        child: CircularProgressIndicator(
-                          value: _progress / 100,
-                        ),
-                      ),
-                    ),
-                  ),
-                ):SizedBox.shrink(),
+                StatefulBuilder(builder: (context, _setState) {
+                  progressStateSetter = _setState;
+                  return _progress != 100
+                      ? Positioned(
+                          child: GestureDetector(
+                            behavior: HitTestBehavior.opaque,
+                            child: Center(
+                              child: Container(
+                                width: width(100),
+                                height: width(100),
+                                child: CircularProgressIndicator(
+                                  value: _progress / 100,
+                                ),
+                              ),
+                            ),
+                          ),
+                        )
+                      : SizedBox.shrink();
+                }),
               ],
             ),
           );
