@@ -1,17 +1,17 @@
 import 'dart:async';
+import 'package:fluro/fluro.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_play/routerPath.dart';
-import 'package:flutter_screenutil/flutter_screenutil.dart';
+import 'package:flutter_play/router/path.dart';
 import 'package:provider/provider.dart';
 
-import 'package:flutter_play/router.dart';
+import 'package:flutter_play/router/router.dart';
 import 'package:flutter_play/components/GlobalComponents.dart';
 import 'package:flutter_play/variable.dart';
 import 'package:flutter_play/store/model.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
-class SplashBanner extends StatefulWidget{
+class SplashBanner extends StatefulWidget {
   @override
   State<StatefulWidget> createState() {
     return SplashBannerState();
@@ -19,8 +19,7 @@ class SplashBanner extends StatefulWidget{
 }
 
 class SplashBannerState extends State<SplashBanner> {
-
-  Timer _timer;
+  Timer? _timer;
   int count = 3;
 
   @override
@@ -66,7 +65,7 @@ class SplashBannerState extends State<SplashBanner> {
 
   @override
   void initState() {
-    if (_timer!=null) _timer.cancel();
+    if (_timer != null) _timer!.cancel();
     _timer = Timer.periodic(Duration(seconds: 1), (Timer timer) {
       setState(() {
         count--;
@@ -78,16 +77,13 @@ class SplashBannerState extends State<SplashBanner> {
     super.initState();
   }
 
-
   _goEntry() async {
-    if (_timer!=null) _timer.cancel();
+    if (_timer != null) _timer!.cancel();
     SharedPreferences sp = await SharedPreferences.getInstance();
     sp.setBool("splash", true);
-    RouterManager.router.navigateTo(
-      context,
-      EntryPage.name,
-      clearStack: true,
-      transitionDuration: Duration(milliseconds: 500),
-    );
+    RouterManager.router!.navigateTo(context, EntryPage.name,
+        clearStack: true,
+        transitionDuration: Duration(milliseconds: 500),
+        transition: TransitionType.fadeIn);
   }
 }
