@@ -3,24 +3,16 @@ import 'dart:convert';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:dio/dio.dart';
 import 'package:flutter_inner_drawer/inner_drawer.dart' show InnerDrawerState;
-import 'package:flutter_play/pages/global/scan.dart';
 import 'package:flutter_play/pages/home/search.dart';
-import 'package:flutter_easyrefresh/easy_refresh.dart';
-import 'package:permission_handler/permission_handler.dart';
-import 'package:flutter_screenutil/flutter_screenutil.dart';
 
-import 'package:flutter_play/service.dart';
 import 'package:flutter_play/variable.dart';
-import 'package:flutter_play/utils/utils.dart';
 import 'package:flutter_play/components/GlobalComponents.dart';
-import 'homeComponent.dart' as Home;
 
 class HomePage extends StatefulWidget {
-  HomePage({ this.drawerKey });
+  HomePage({this.drawerKey});
 
-  final GlobalKey<InnerDrawerState> drawerKey;
+  final GlobalKey<InnerDrawerState>? drawerKey;
 
   @override
   createState() => HomePageState();
@@ -28,7 +20,7 @@ class HomePage extends StatefulWidget {
 
 class HomePageState extends State<HomePage> with AutomaticKeepAliveClientMixin {
   bool _hideInputClear = true;
-  FocusNode _focusNode;
+  FocusNode? _focusNode;
   final _inputController = MyTextEditingController();
   String get query => _inputController.text;
   set query(String value) {
@@ -36,11 +28,9 @@ class HomePageState extends State<HomePage> with AutomaticKeepAliveClientMixin {
     _inputController.text = value;
   }
 
-  final _refreshController = EasyRefreshController();
   final _scrollController = ScrollController();
   bool _hideGoTop = true;
   double _goTopOpacity = 0;
-
 
   // data
   List<dynamic> bannerList = [];
@@ -58,22 +48,14 @@ class HomePageState extends State<HomePage> with AutomaticKeepAliveClientMixin {
         centerTitle: false,
         titleSpacing: 0,
         toolbarHeight: 50,
-        actions: <Widget>[
-          IconButton(
-            icon: Icon(const IconData(0xe610, fontFamily: 'iconfont')),
-            onPressed: (){
-              _focusNode?.unfocus();
-              goScan();
-            },
-          )
-        ],
         leading: Builder(
           builder: (BuildContext context) {
             return IconButton(
               icon: Icon(Icons.menu),
               onPressed: () {
                 _focusNode?.unfocus();
-                if (widget.drawerKey!=null) widget.drawerKey.currentState.open();
+                if (widget.drawerKey != null)
+                  widget.drawerKey!.currentState?.open();
               },
             );
           },
@@ -87,8 +69,7 @@ class HomePageState extends State<HomePage> with AutomaticKeepAliveClientMixin {
               style: TextStyle(
                   fontSize: width(30),
                   textBaseline: TextBaseline.alphabetic,
-                  color: Colors.black
-              ),
+                  color: Colors.black),
               focusNode: _focusNode,
               controller: _inputController,
               cursorColor: Theme.of(context).primaryColor,
@@ -123,80 +104,13 @@ class HomePageState extends State<HomePage> with AutomaticKeepAliveClientMixin {
           ),
         ),
       ),
-      body: GestureDetector(
-        onTap: () {
-          _focusNode?.unfocus();
-        },
-        child: Stack(
-          children: <Widget>[
-            CupertinoScrollbar(
-              child: EasyRefresh(
-                header: ClassicalHeader(
-                  refreshText: '下拉刷新数据',
-                  refreshReadyText: '释放刷新',
-                  refreshingText: '刷新中',
-                  refreshedText: '刷新数据成功',
-                  refreshFailedText: '刷新数据失败',
-                ),
-                footer: ClassicalFooter(
-                  loadText: '上拉加载',
-                  loadingText: '加载中',
-                  loadedText: '加载完成',
-                  loadReadyText: '松开加载更多',
-                  noMoreText: '没有更多数据',
-                  loadFailedText: '加载失败，点击重试',
-                ),
-                controller: _refreshController,
-                onRefresh: _refresh,
-                onLoad: _loadMore,
-                enableControlFinishRefresh: true,
-                enableControlFinishLoad: true,
-                child: ListView(
-                  physics: BouncingScrollPhysics(),
-                  controller: _scrollController,
-                  children: <Widget>[
-                    Home.Banner(bannerList),
-                    Home.Tag(),
-                    Home.RecommendMusicList(recommendMusicList),
-                  ],
-                ),
-              ),
-            ),
-            Positioned(
-              child: Offstage(
-                offstage: _hideGoTop,
-                child: AnimatedOpacity(
-                  opacity: _goTopOpacity,
-                  duration: Duration(milliseconds: 600),
-                  child: Container(
-                    width: 40,
-                    height: 40,
-                    decoration: BoxDecoration(
-                      borderRadius: BorderRadius.all(Radius.circular(40)),
-                      color: Colors.black,
-                    ),
-                    child: Center(
-                      child: IconButton(
-                        icon: Icon(Icons.arrow_upward, color: Colors.white),
-                        onPressed: _goTop,
-                      ),
-                    ),
-                  ),
-                ),
-              ),
-              bottom: 10,
-              right: 10,
-            ),
-          ],
-        ),
-      ),
+      body: ListView(),
     );
   }
 
   @override
   void initState() {
     _focusNode = FocusNode(debugLabel: "search");
-    _refresh();
     _inputController.addListener(_inputChange);
     _scrollController.addListener(_scrollListen);
     super.initState();
@@ -209,37 +123,16 @@ class HomePageState extends State<HomePage> with AutomaticKeepAliveClientMixin {
 
   @override
   void dispose() {
-    _focusNode.dispose();
-    _refreshController.dispose();
+    _focusNode?.dispose();
     _inputController.removeListener(_inputChange);
     _scrollController.removeListener(_scrollListen);
     super.dispose();
   }
 
-  void _refresh() async {
-    try {
-      // banner
-      bannerList = await Service.getHomeBanner();
-      recommendMusicList = await Service.getRecommendMusicList();
-      setState(() {
-        _refreshController.finishRefresh(success: true);
-      });
-    } on DioError catch(e) {
-      setState(() {
-        _refreshController.finishRefresh(success: false);
-      });
-    }
-  }
-
-  void _loadMore() async {
-   await Future.delayed(Duration(seconds: 2));
-    setState(() {
-      _refreshController.finishLoad(success: true);
-    });
-  }
-
   void _submit(value) {
-    Navigator.of(context).pushNamed('${SearchPage.name}?text=${Uri.encodeQueryComponent(value)}',);
+    Navigator.of(context).pushNamed(
+      '${SearchPage.name}?text=${Uri.encodeQueryComponent(value)}',
+    );
   }
 
   void _inputChange() {
@@ -284,28 +177,7 @@ class HomePageState extends State<HomePage> with AutomaticKeepAliveClientMixin {
   }
 
   void _goTop() {
-    _scrollController.animateTo(0, duration: Duration(milliseconds: 600), curve: Curves.linear);
-  }
-
-  void goScan() async {
-    PermissionStatus permissionStatus = (await Permission.camera.status);
-    bool hasPermission = permissionStatus==PermissionStatus.granted;
-    bool unknown = (permissionStatus.isUndetermined) || (permissionStatus.isDenied);
-    if (unknown) {
-      hasPermission = await Permission.camera.request().isGranted;
-    }
-    if (hasPermission) {
-      // Navigator.of(context).pushNamed(ScanPage.name);
-    } else{
-      MyDialog(
-        context: context,
-        title: '扫码提示',
-        content: '扫码需要允许相机权限',
-        confirmText: '前往设置',
-        onConfirm: () {
-          openAppSettings();
-        }
-      );
-    }
+    _scrollController.animateTo(0,
+        duration: Duration(milliseconds: 600), curve: Curves.linear);
   }
 }
