@@ -2,19 +2,17 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'dart:ui';
 import 'package:crypto/crypto.dart';
-import 'package:convert/convert.dart';
 import 'dart:convert';
-
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-
+export './iconfont.dart';
 
 GlobalKey rootKey = GlobalKey();
 
 GlobalKey<NavigatorState> navigatorKey = GlobalKey<NavigatorState>();
 
-OverlayState globalOverlayState;
+OverlayState? globalOverlayState;
 
-BuildContext globalContext;
+BuildContext? globalContext;
 
 double statusBarHeight = MediaQueryData.fromWindow(window).padding.top;
 
@@ -26,7 +24,7 @@ double vw = MediaQueryData.fromWindow(window).size.width;
 
 double vh = MediaQueryData.fromWindow(window).size.height;
 
-bool isSystemDark = MediaQueryData.fromWindow(window).platformBrightness==Brightness.dark;
+bool isSystemDark = MediaQueryData.fromWindow(window).platformBrightness == Brightness.dark;
 
 final List<ThemeMode> themeModeList = [ThemeMode.system, ThemeMode.light, ThemeMode.dark];
 
@@ -48,11 +46,11 @@ String durationToTime(Duration duration) {
   int h = duration.inHours;
   int m = duration.inMinutes.remainder(60);
   int s = duration.inSeconds.remainder(60);
-  return '${h>0?h.toString()+':':''}${m<10?'0'+m.toString():m}:${s<10?'0'+s.toString():s}';
+  return '${h > 0 ? h.toString() + ':' : ''}${m < 10 ? '0' + m.toString() : m}:${s < 10 ? '0' + s.toString() : s}';
 }
 
 String generateToken(String appid, String secretKey, String shortid) {
-  int expiredTime = DateTime(2019, 7, 13).millisecondsSinceEpoch~/100;
+  int expiredTime = DateTime(2019, 7, 13).millisecondsSinceEpoch ~/ 100;
   final rawMask = '${appid}_${expiredTime}_$secretKey';
   final mask = Hmac(sha1, utf8.encode(secretKey)).convert(utf8.encode(rawMask));
   final token = '$shortid:$expiredTime:$mask';
@@ -67,120 +65,17 @@ String generateToken(String appid, String secretKey, String shortid) {
 double width(num w) {
   return ScreenUtil().setWidth(w).toDouble();
 }
+
 double height(num h) {
   return ScreenUtil().setHeight(h).toDouble();
 }
 
-class MyTheme {
-  static ThemeData light(Color themeColor) {
-    return ThemeData(
-      appBarTheme: AppBarTheme(
-        elevation: 0,
-        brightness: Brightness.light,
-        textTheme: TextTheme(
-          bodyText1: TextStyle(
-            decoration: TextDecoration.none,
-            color: Colors.black,
-          ),
-          button: TextStyle(
-            decoration: TextDecoration.none,
-            color: Colors.black,
-          ),
-        ),
-        iconTheme: IconThemeData(color: Colors.black),
-        actionsIconTheme: IconThemeData(color: Colors.black),
-        color: themeColor,
-      ),
-      brightness: Brightness.light,
-      primaryColorBrightness: Brightness.light,
-      primaryColor: themeColor,
-      accentColor: themeColor,
-      cursorColor: themeColor,
-      textSelectionColor: themeColor,
-      scaffoldBackgroundColor: Colors.white,
-      textTheme: TextTheme(
-        bodyText1: TextStyle(
-          decoration: TextDecoration.none,
-          color: Colors.black,
-        ),
-        button: TextStyle(
-          decoration: TextDecoration.none,
-          color: Colors.black,
-        ),
-      ),
-      buttonTheme: ButtonThemeData(
-        buttonColor: themeColor,
-      ),
-      iconTheme: IconThemeData(
-        color: Colors.black
-      ),
-      cupertinoOverrideTheme: CupertinoThemeData(
-        brightness: Brightness.light,
-        primaryColor: themeColor,
-        scaffoldBackgroundColor: Colors.white,
-      ),
-    );
-  }
-
-  static ThemeData dark(Color themeColor) {
-    return ThemeData(
-      appBarTheme: AppBarTheme(
-        elevation: 0,
-        brightness: Brightness.dark,
-        textTheme: TextTheme(
-          bodyText1: TextStyle(
-            decoration: TextDecoration.none,
-            color: Colors.white,
-            fontFamily: '-apple-system-font,Helvetica Neue,Helvetica,sans-serif',
-          ),
-          button: TextStyle(
-            decoration: TextDecoration.none,
-            color: Colors.white,
-            fontFamily: '-apple-system-font,Helvetica Neue,Helvetica,sans-serif',
-          ),
-        ),
-        iconTheme: IconThemeData(color: Colors.white),
-        actionsIconTheme: IconThemeData(color: Colors.white),
-        color: themeColor,
-      ),
-      brightness: Brightness.dark,
-      primaryColorBrightness: Brightness.dark,
-      primaryColor: themeColor,
-      accentColor: themeColor,
-      cursorColor: themeColor,
-      textSelectionColor: themeColor,
-      scaffoldBackgroundColor: Colors.black,
-      textTheme: TextTheme(
-        bodyText1: TextStyle(
-          decoration: TextDecoration.none,
-          color: Colors.white,
-        ),
-        button: TextStyle(
-          decoration: TextDecoration.none,
-          color: Colors.white,
-        ),
-      ),
-      buttonTheme: ButtonThemeData(
-        buttonColor: themeColor,
-      ),
-      iconTheme: IconThemeData(
-        color: Colors.white
-      ),
-      cupertinoOverrideTheme: CupertinoThemeData(
-        brightness: Brightness.dark,
-        primaryColor: themeColor,
-        scaffoldBackgroundColor: Colors.black,
-      ),
-    );
-  }
-}
-
 void uiInit(BuildContext context, BoxConstraints boxConstraints) {
-  final BuildContext _context = context ?? globalContext;
+  final BuildContext _context = context;
 
   final Orientation currentOrientation = MediaQuery.of(_context).orientation;
 
-  if (currentOrientation==Orientation.portrait) {
+  if (currentOrientation == Orientation.portrait) {
     vw = MediaQuery.of(_context).size.width;
 
     vh = MediaQuery.of(_context).size.height;
@@ -192,12 +87,12 @@ void uiInit(BuildContext context, BoxConstraints boxConstraints) {
 
   dpr = MediaQuery.of(_context).devicePixelRatio;
 
-  isSystemDark = MediaQuery.of(_context).platformBrightness==Brightness.dark;
+  isSystemDark = MediaQuery.of(_context).platformBrightness == Brightness.dark;
 
   //
   globalContext = context;
 
   globalOverlayState = Overlay.of(context);
 
-  ScreenUtil.init(boxConstraints, designSize: Size(750, 1134));
+  ScreenUtil.init(boxConstraints, Orientation.portrait, designSize: Size(750, 1134));
 }
