@@ -3,9 +3,7 @@ import 'package:flutter/material.dart';
 import '../variable.dart';
 
 class MyToast extends StatefulWidget {
-  MyToast({
-    this.msg,
-  });
+  MyToast(this.msg);
 
   final String msg;
 
@@ -23,11 +21,8 @@ class _MyToastState extends State<MyToast> {
           maxWidth: width(500),
         ),
         color: Color.fromRGBO(0, 0, 0, .9),
-        padding: EdgeInsets.only(
-            left: width(30),
-            right: width(30),
-            top: width(20),
-            bottom: width(20)),
+        padding:
+            EdgeInsets.only(left: width(30), right: width(30), top: width(20), bottom: width(20)),
         child: Text(
           widget.msg,
           maxLines: 3,
