@@ -1,11 +1,12 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_play/generated/l10n.dart';
 import 'package:provider/provider.dart';
 
 import 'package:flutter_play/store/model.dart';
 import 'package:flutter_play/components/GlobalComponents.dart';
-import 'package:flutter_play/pages/global/imagePreview.dart';
 import 'package:flutter_play/variable.dart';
+import 'package:pull_to_refresh/pull_to_refresh.dart';
 
 class UserCenter extends StatefulWidget {
   static const String title = 'user';
@@ -15,67 +16,105 @@ class UserCenter extends StatefulWidget {
   createState() => UserCenterState();
 }
 
-class UserCenterState extends State<UserCenter> with AutomaticKeepAliveClientMixin,WidgetsBindingObserver {
+class UserCenterState extends State<UserCenter>
+    with AutomaticKeepAliveClientMixin {
   @override
   bool get wantKeepAlive => true;
 
+  final _refreshController = RefreshController();
+
   Widget build(BuildContext context) {
     super.build(context);
-    return Scaffold(
-      appBar: PreferredSize(
-        preferredSize: Size.fromHeight(50),
-        child: AppBar(
-          title: Text('个人中心'),
-        ),
-      ),
-      body: Container(
-        child: ListView(
-          physics: ClampingScrollPhysics(),
-          children: <Widget>[
-            Column(
-              children: <Widget>[
-                MyImage(
-                  'https://cdn.chavesgu.com/avatar.jpg',
-                  width: 150,
-                  height: 150,
-                  shape: BoxShape.circle,
-                  preview: true,
-                ),
-                MyImage('assets/images/avatar.png'),
-                Selector<UserModel, String>(
-                  selector: (context, model) => model.title,
-                  builder: (context, value, child) {
-                    return Text(value);
-                  },
-                )
-              ],
+    return SmartRefresher(
+      header: MaterialClassicHeader(),
+      controller: _refreshController,
+      onRefresh: _refresh,
+      child: CustomScrollView(
+        slivers: [
+          SliverAppBar(
+            title: Selector<GlobalModel, Locale>(
+              builder: (context, locale, child) {
+                return Text(S.of(context).user);
+              },
+              selector: (context, model) => model.lang!,
             ),
-          ],
-        ),
+            centerTitle: true,
+            pinned: true,
+            toolbarHeight: 50,
+            expandedHeight: 150.0,
+            flexibleSpace: FlexibleSpaceBar(
+              background: MyImage(
+                'assets/images/bilibili.jpg',
+              ),
+            ),
+          ),
+          SliverList(
+            delegate: SliverChildListDelegate([
+              Row(
+                children: [
+                  MyImage(
+                    'https://cdn.chavesgu.com/avatar.jpg',
+                    width: 50,
+                    height: 50,
+                    shape: BoxShape.circle,
+                    preview: true,
+                  ),
+                  Selector<UserModel, String>(
+                    selector: (context, model) => model.title,
+                    builder: (context, value, child) {
+                      return Text(value);
+                    },
+                  ),
+                ],
+              ),
+              Row(
+                children: [
+                  ElevatedButton(
+                    child: Text('切换en'),
+                    onPressed: () {
+                      context
+                          .read<GlobalModel>()
+                          .changeLocale(Locale.fromSubtags(languageCode: 'en'));
+                    },
+                  ),
+                  ElevatedButton(
+                    child: Text('切换zh'),
+                    onPressed: () {
+                      context.read<GlobalModel>().changeLocale(
+                          Locale.fromSubtags(
+                              languageCode: 'zh', scriptCode: 'Hans'));
+                    },
+                  ),
+                ],
+              ),
+              ...buildList(),
+            ]),
+          ),
+        ],
       ),
     );
   }
 
-  void _scale(String url, TapUpDetails details) {
-    double _x = (vw/2 - details.globalPosition.dx)/(vw/2);
-    double _y = (vh/2 - details.globalPosition.dy)/(vh/2);
-    Navigator.push(
-      context,
-      PageRouteBuilder(
-        pageBuilder: (BuildContext context, Animation animation, Animation secondaryAnimation) {
-          return ScaleTransition(
-            scale: animation,
-            alignment: Alignment(
-              -_x,
-              -_y
-            ),
-            child: ImagePreview(
-              imageList: [url, url],
-            ),
-          );
-        },
-        transitionDuration: Duration(milliseconds: 150),
-      )
-    );
+  @override
+  void dispose() {
+    _refreshController.dispose();
+    super.dispose();
+  }
+
+  List<Widget> buildList() {
+    final List<Widget> list = [];
+    for (int i = 0; i < 20; i++) {
+      list.add(ListTile(
+        title: Text('chaves'),
+        subtitle: Text('$i'),
+        trailing: Icon(IconFont.right_arrow),
+      ));
+    }
+    return list;
+  }
+
+  void _refresh() async {
+    await Future.delayed(Duration(seconds: 2));
+    _refreshController.refreshCompleted();
   }
 }
