@@ -1,15 +1,14 @@
 library router_path;
 
 export 'package:flutter_play/pages/entry.dart';
-export 'package:flutter_play/pages/global/musicListDetail.dart';
 export 'package:flutter_play/pages/global/WebView.dart';
 export 'package:flutter_play/pages/home/search.dart';
 export 'package:flutter_play/pages/global/scan.dart';
-export 'package:flutter_play/pages/global/topList.dart';
-export 'package:flutter_play/pages/drawer/about.dart';
 export 'package:flutter_play/pages/drawer/setting.dart';
 export 'package:flutter_play/pages/demo/testFixed.dart';
-export 'package:flutter_play/pages/demo/testGesture.dart';
 export 'package:flutter_play/pages/demo/pdf.dart';
 export 'package:flutter_play/pages/demo/canvas1.dart';
 export 'package:flutter_play/pages/demo/canvas2.dart';
+export 'package:flutter_play/pages/demo/chart.dart';
+export 'package:flutter_play/pages/media/videoDetail.dart';
+export 'package:flutter_play/pages/media/audioDetail.dart';
